Creates default Gadgetron config and set env
it creates a default Gadgetron configuration file based on the
gadgetron.xml.example ONLY IF the gadgetron.xml file does not exist.
Users will be able to run Gadgetron after build without modifying the file.

Appends to .bashrc only if not already done. It seems this is not wanted,
under discussion.