set required version for some dependencies when using system versions
When using the SYSTEM version of a dependency, require HDF5 1.8 (1.8.17 on
Windows), Boost 1.36.0 and SWIG 2. scope for more.

Most/all of these should be caught by later on by the SIRF CMake file
etc, but putting them here means that the user will see failure at
initial configure time, as opposed to somewhere half-way the build.

It also fixes a problem on Windows that ISMRMRD doesn't currently check
for 1.8.17.