Mark as superbuild and other fixes for badly supported packages (#148)
* Remove setting of FFTW3_ROOT_DIR CMake variable

Remove as it's ignored (see #147).
Note that it was wrongly spelled in External_Gadgetron.cmake

* pass SWIG_EXECUTABLE to all projects

* correct use of mark_as_superbuild

Use `ALL_PROJECTS` as opposed to `PROJECTS ALL_PROJECTS`. This
seems to have the same effect though. See #70.

* small corrections to FindACE.cmake

- Don't set ACE_DIR with makr_as_superbuild. Should have been ACE_ROOT, but
it isn't used anyway
- correct variable names

* some clean-up of Armadillo

extra comments and remove mark_as_superbuild as not used.

* correct setting of SWIG variables, adding SWIG_EXECUTABLE to the cache

This also allows people to change the executable if they want.

* correct diagnostics when using system ACE