try to fix travis osx boost1.69/gadgetron (#247)
As of boost 1.69, boost.Process is broken. https://stackoverflow.com/a/54912043

To get travis working, use `brew install boost` and then manually edit headers to remove `::` from all occurences of `::sigtimedwait`, `::sigaddset` and `::sigemptyset`.

This is a temporary fix to `.travis.yml` that can hopefully be deleted in the future.