docker: introduce extra user for building
To avoid having root owning too many files, and to prepare for jupyterhub
(which doesn't want root), we create the user* files using a new user
(currently defaulting to be called jovyan).

entrypoint.sh and sirf-compose will still create sirfuser such that it can
read/write to the mounted volume.

This way, we don't have to change permissions/ownership if most of the files, saving
time. However, it does mean that conda needs to be run by jovyan, which is undesirable.

[ci skip]