VM: UPDATE.sh no longer runs configure_gnome.sh
This seems dangerous as it could override user settings.
In addition, it allows us to simplify the bootstrap script a bit.
This also means that INSTALL*sh scripts appropriate to the desired version are used
(before they were the ones on master).

I had to split the optional updates then however, as apt needs to be run
before we can find the python executable. Similarly, python updates have
to be run before we can find cython.

Fixes #739